Extract internal code to base class
