+++ sdk/ruby/lib/dagger.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module Dagger
+  require_relative "dagger/base"
   require_relative "dagger/client"
   require_relative "dagger/connect"
   require_relative "dagger/version"
